+++ source/Delphi.Forms.Info.pas
@@ -3,7 +3,7 @@ unit Delphi.Forms.Info;
 interface
 
 const
-  Version = '0.8.43.0';
+  Version = '0.8.44.0';
 
 implementation